Move mutex creation out of line.
Move the mutex create and destroy out of the header file so that
we can use static mutex attributes without exposing another global
variable.
