+++ lib/ts/ink_mutex.cc
@@ -29,19 +29,46 @@
 
 ink_mutex __global_death = PTHREAD_MUTEX_INITIALIZER;
 
-x_pthread_mutexattr_t::x_pthread_mutexattr_t()
+class x_pthread_mutexattr_t
 {
-  pthread_mutexattr_init(&attr);
+public:
+  x_pthread_mutexattr_t()
+  {
+    pthread_mutexattr_init(&attr);
 #ifndef POSIX_THREAD_10031c
-  pthread_mutexattr_setpshared(&attr, PTHREAD_PROCESS_SHARED);
+    pthread_mutexattr_setpshared(&attr, PTHREAD_PROCESS_SHARED);
 #endif
 
 #if DEBUG && HAVE_PTHREAD_MUTEXATTR_SETTYPE
-  pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_ERRORCHECK);
+    pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_ERRORCHECK);
 #endif
+  }
+
+  ~x_pthread_mutexattr_t() { pthread_mutexattr_destroy(&attr); }
+
+  pthread_mutexattr_t attr;
+};
+
+static x_pthread_mutexattr_t attr;
+
+void
+ink_mutex_init(ink_mutex *m, const char * /* name */)
+{
+  int error;
+
+  error = pthread_mutex_init(m, &attr.attr);
+  if (unlikely(error != 0)) {
+    ink_abort("pthread_mutex_init(%p) failed: %s (%d)", m, strerror(error), error);
+  }
 }
 
-x_pthread_mutexattr_t::~x_pthread_mutexattr_t()
+void
+ink_mutex_destroy(ink_mutex *m)
 {
-  pthread_mutexattr_destroy(&attr);
+  int error;
+
+  error = pthread_mutex_destroy(m);
+  if (unlikely(error != 0)) {
+    ink_abort("pthread_mutex_destroy(%p) failed: %s (%d)", m, strerror(error), error);
+  }
 }
